Changed ids to Long. Added new daos using jpa, woring for tag.

+++ core/src/main/java/com/epam/esm/dto/TagDto.java
@@ -5,16 +5,14 @@ import com.fasterxml.jackson.annotation.JsonProperty;
 import lombok.EqualsAndHashCode;
 import lombok.Getter;
 
-import java.math.BigInteger;
-
 @Getter
 @EqualsAndHashCode
 public class TagDto {
-    private final BigInteger id;
+    private final Long id;
     private final String name;
 
     @JsonCreator(mode = JsonCreator.Mode.PROPERTIES)
-    public TagDto(@JsonProperty("id") BigInteger id,
+    public TagDto(@JsonProperty("id") Long id,
                   @JsonProperty("name") String name) {
         this.id = id;
         this.name = name;
